+++ lib/src/actions/auth.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:redux_thunk_boilerplate/src/actions/global.dart';
 import 'package:redux/redux.dart';
 
+import 'package:redux_thunk_boilerplate/src/actions/global.dart' as global_actions;
 import 'package:redux_thunk_boilerplate/src/models/app_state.dart';
 import 'package:redux_thunk_boilerplate/src/models/user.dart';
 import 'package:redux_thunk_boilerplate/src/ui/screens/dashboard.dart';
@@ -18,20 +18,20 @@ class LogoutSuccess {}
 
 final Function onLogin = (BuildContext context) {
   return (Store<AppState> store) {
-    store.dispatch(IncrementLoadingRequest());
+    store.dispatch(global_actions.IncrementLoadingRequest());
     // TODO: login request
     store.dispatch(LoginSuccess('some_token', User(id: 'some_id', email: 'test@testmail.com', name: 'Test')));
-    store.dispatch(DecrementLoadingRequest());
     Navigator.of(context).pushNamedAndRemoveUntil(DashboardScreen.route, (_) => false);
+    store.dispatch(global_actions.DecrementLoadingRequest());
   };
 };
 
 final Function onLogout = (BuildContext context) {
   return (Store<AppState> store) {
-    store.dispatch(IncrementLoadingRequest());
+    store.dispatch(global_actions.IncrementLoadingRequest());
     // TODO: logout request
-    store.dispatch(LogoutSuccess());
-    store.dispatch(DecrementLoadingRequest());
+    store.dispatch(global_actions.Reset());
     Navigator.of(context).pushNamedAndRemoveUntil(LoginScreen.route, (_) => false);
+    store.dispatch(global_actions.DecrementLoadingRequest());
   };
 };