+++ lib/src/reducers/global.dart
@@ -1,19 +1,19 @@
 import 'package:redux/redux.dart';
 
-import 'package:redux_thunk_boilerplate/src/actions/global.dart' as actions;
+import 'package:redux_thunk_boilerplate/src/actions/global.dart' as global_actions;
 import 'package:redux_thunk_boilerplate/src/models/global_state.dart';
 
 Reducer<GlobalState> globalReducer = combineReducers([
-  TypedReducer<GlobalState, actions.IncrementLoadingRequest>(incrementLoadingReducer),
-  TypedReducer<GlobalState, actions.DecrementLoadingRequest>(decrementLoadingReducer),
+  TypedReducer<GlobalState, global_actions.IncrementLoadingRequest>(_incrementLoadingReducer),
+  TypedReducer<GlobalState, global_actions.DecrementLoadingRequest>(_decrementLoadingReducer),
 ]);
 
-GlobalState incrementLoadingReducer(GlobalState state, actions.IncrementLoadingRequest action) {
+GlobalState _incrementLoadingReducer(GlobalState state, global_actions.IncrementLoadingRequest action) {
   final loading = state.loading + action.amount;
   return GlobalState().copyWith(loading: loading < 0 ? 0 : loading);
 }
 
-GlobalState decrementLoadingReducer(GlobalState state, actions.DecrementLoadingRequest action) {
+GlobalState _decrementLoadingReducer(GlobalState state, global_actions.DecrementLoadingRequest action) {
   final loading = state.loading - action.amount;
   return GlobalState().copyWith(loading: loading < 0 ? 0 : loading);
 }